Move benching into testing.
