+++ testing/bench_test.ts
@@ -1,22 +1,22 @@
-import { test } from "../testing/mod.ts";
-import { bench, runBenchmarks, BenchmarkTimer } from "./mod.ts";
+import { test, runIfMain } from "./mod.ts";
+import { bench, runBenchmarks } from "./bench.ts";
 
-import "example.ts";
+import "./bench_example.ts";
 
 test(async function benching() {
-  bench(function forIncrementX1e9(b: BenchmarkTimer) {
+  bench(function forIncrementX1e9(b) {
     b.start();
     for (let i = 0; i < 1e9; i++);
     b.stop();
   });
 
-  bench(function forDecrementX1e9(b: BenchmarkTimer) {
+  bench(function forDecrementX1e9(b) {
     b.start();
     for (let i = 1e9; i > 0; i--);
     b.stop();
   });
 
-  bench(async function forAwaitFetchDenolandX10(b: BenchmarkTimer) {
+  bench(async function forAwaitFetchDenolandX10(b) {
     b.start();
     for (let i = 0; i < 10; i++) {
       await fetch("https://deno.land/");
@@ -24,7 +24,7 @@ test(async function benching() {
     b.stop();
   });
 
-  bench(async function promiseAllFetchDenolandX10(b: BenchmarkTimer) {
+  bench(async function promiseAllFetchDenolandX10(b) {
     const urls = new Array(10).fill("https://deno.land/");
     b.start();
     await Promise.all(urls.map((denoland: string) => fetch(denoland)));
@@ -34,17 +34,19 @@ test(async function benching() {
   bench({
     name: "runs100ForIncrementX1e6",
     runs: 100,
-    func(b: BenchmarkTimer) {
+    func(b) {
       b.start();
       for (let i = 0; i < 1e6; i++);
       b.stop();
     }
   });
 
-  bench(function throwing(b: BenchmarkTimer) {
+  bench(function throwing(b) {
     b.start();
     // Throws bc the timer's stop method is never called
   });
 
   await runBenchmarks({ skip: /throw/ });
 });
+
+runIfMain(import.meta);
